Refactored common testing code into debug folder, added test using common code
The js files in debug folders are loaded in when running sc-server.  But they
will be ignored when it is built.  This appears to be the best place to put common
test code.  js files in the test folders are treated as tests and are wrapped in anon
functions.  Additionally someone can run your test js file directly so you can't
count on it including other files.

The new test excercises the common code but doesn't use it yet.