cut off recursion in notifyStatus and added async retrieveRecord
The previous notifyStatus could cause an infinite loop if the callback func caused
status to change.  This is fixed by removing the observer before the callback is called.

retrieveRecord now works asynchronously.  Multiple tests had to be updated to continue working with this
change.  This is because multiple tests are calling 'find' with a record type and id.