Added async remove code and test
In addition to this async code itself there is a monkey patch of the qunit stop metthod so it can run code when the test timeout.
Also I found that the async callbacks need to return YES if the callback handled the response.  So I updated all of the callbacks to do this correctly.