Fixed create bug and added some new support methods
The create code wasn't setting the id of the created object correctly.  This showed up when an object was created and then it was updated right after.  The new support methods will be used to help apps setup persevere.  Additionally this commit contains some formating changes so now things are generally indented with 2 space.  I'm removing tabs as I work with them.